remove diode through-holes, move diode pads to led-hole
through hole diodes obv. don't fit on the side the switches are on,
and are taller than the feet on the other side.
we're looking for thinness here, so through-hole diodes are out

moved the diode pads so the (smd) diodes sit in the void choc minis have
for LEDs

also using the adjacent foot through-holes of the switch as a via for
the diode pads
those feet aren't electrically connected to anything so that should be
fine